♻️ Refactor : Bedge추가, css수정
