+++ src/components/PostComponent/PostComponent.css
@@ -14,7 +14,7 @@
 .post-card .title {
   font-family: 'NotoSansKR';
 
-  font-size: 20px;
+  font-size: 24px;
   text-align: left; 
   width: 420px;
   padding-top: 20px;
@@ -25,9 +25,9 @@
   display:flex;
   flex-direction: row;
   justify-content:space-between ;
-  font-size: 12px;
+  font-size: 13px;
   color: #7d7d7d;
-  margin: 4px 0px;
+  margin: 10px 0px;
   width: 420px;
 }
 
@@ -45,5 +45,22 @@
   margin: 16px 0 16px;
 }
 
+.category_left {
+  display: flex;
+  flex-direction:row ;
+  gap:4px;
+  align-items: center;
+}
 
+.category_right {
+  display: flex;
+  flex-direction:row ;
+  align-items: center;
+  gap:4px;
+}
 
+.right_icon{
+  display: flex;
+  align-items: center;
+  gap:1px;
+}
